backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-docx
PyPDF2
langchain
langchain-openai
langchain-core
python-dotenv
aiofiles
email-validator
jinja2
pillow
